Move tokenState file to utils
